% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dataExtract}
\alias{dataExtract}
\title{get data summary}
\usage{
dataExtract(mydf, pattern)
}
\arguments{
\item{mydf}{data to extract}

\item{pattern}{pattern to choose from}
}
\description{
get data summary
}
