% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pattern_plot}
\alias{pattern_plot}
\title{pattern plot}
\usage{
pattern_plot(data, pattern)
}
\arguments{
\item{data}{data for plot}

\item{pattern}{pattern (conforms to PERL style)}
}
\value{
ggplot object
}
\description{
pattern plot
}
